Allow for earlier versions of pyyaml (#82)
## 💌 Description

Requiring pyyaml>=6 makes it impossible to install this tool if other
dependencies in your project, such as awscli, require pyyaml<6.

Giving this dependency more flexibility will allow me to use it in my
project. I assume it will not break the behavior of the tool. Please let
me know if I'm wrong or if this project is directly depending on
features only available in Pyyaml 6.

Thanks for making this!

Co-authored-by: Dariusz Porowski <3431813+DariuszPorowski@users.noreply.github.com>